server/bedrock/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# AWS and Bedrock
boto3>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0

# Data validation and utilities
pydantic>=2.0.0
ulid-py>=1.1.0

# AWS Lambda Powertools (for metrics and logging)
aws-lambda-powertools>=2.0.0

# Optional: For local development and testing
moto>=5.0.0  # Mock AWS services
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For testing FastAPI
